Name validation on Project.define
Project.define makes sure that the name to be defined is not
already being used for a task.
If so, we fail telling the user not to use a task name for his project,
without this validation buildr would still fail but with a Rake error
message that is not so helpful.


git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/buildr/trunk@647023 13f79535-47bb-0310-9956-ffa450edef68