Fixed test property setting when in a subproject.  (BUILDR-126's fix was incomplete)
Each call to test.options[:properties] was returning a different hash when used in a subproject.  The problem was that the block given to the OpenObject's underlying hash was not setting the hash itself, but only returning a clone of the parent's property.

The ruby documentation is quite clear on this issue:
"Hash.new {|hash, key| block } => aHash
<snip>
If a block is specified, it will be called with the hash object and the key, and should return the default value.  It is the block‘s responsibility to store the value in the hash if required."


git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/buildr/trunk@701831 13f79535-47bb-0310-9956-ffa450edef68